metadata working, featured on save only, are you sure delete, obstructor

+++ src/components/metadata/DeleteControl.js
@@ -22,19 +22,48 @@ import React from 'react'
 export default class DeleteControl extends React.Component {
   constructor() {
     super()
+    this.state = {
+      deleteWasClickedOnce: false
+    }
     this._handleClick = this._handleClick.bind(this)
+    this._handleNoClick = this._handleNoClick.bind(this)
     this.render = this.render.bind(this)
   }
   _handleClick(event) {
     event.preventDefault()
-    this.props.deleteItem(this.props.item.get('id'))
+    if (this.state.deleteWasClickedOnce) {
+      this.props.deleteItem(this.props.item.get('id'))
+    }
+    else {
+      this.setState({
+        deleteWasClickedOnce: true
+      })
+    }
+  }
+  _handleNoClick(event) {
+    event.preventDefault()
+    this.setState({
+      deleteWasClickedOnce: false
+    })
   }
   render() {
     if (!this.props.userIsOwner) {
       return null
     }
+    if (this.state.deleteWasClickedOnce) {
+      return (
+        <div className='delete-control are-you-sure'>
+          <span>Are you sure?</span>
+          <a href='#' onClick={this._handleClick} tabIndex='-1'>Yes</a>
+          <a href='#' onClick={this._handleNoClick} tabIndex='-1'>No</a>
+        </div>
+      )
+    }
     return (
-      <a className='delete-control' href='#' onClick={this._handleClick}>
+      <a className='delete-control' 
+         href='#' 
+         onClick={this._handleClick} 
+         tabIndex='-1'>
         Delete
       </a>
     )
